Extract tool-result capability helpers
Move the new marker and restricted-list helpers out of already-large agent files so the implementation stays within repository size budgets without changing behavior or repolint baselines.
